requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-lru==1.0.3
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord==2.0.0
discord.py==2.0.1
google-search-results==2.4.1
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
tweepy==4.10.0
typing_extensions==4.3.0
update-checker==0.18.0
urllib3==1.26.11
websocket-client==1.3.3
yarl==1.7.2
